requirements.txt
astroid==2.9.3
autopep8==1.6.0
click==8.0.3
Flask==2.0.2
gunicorn==20.1.0
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
mccabe==0.6.1
mysql-connector-python==8.0.28
platformdirs==2.4.1
protobuf==3.19.4
pycodestyle==2.8.0
pylint==2.12.2
toml==0.10.2
typing_extensions==4.0.1
Werkzeug==2.0.2
wrapt==1.13.3
